halide2isl: handle non-affine subexpressions
51b9aa8ef814b78d086bd80a1b83d1d661a7d7c2 modified the conversion from
Halide Exprs to isl affs to add some basic support for min/max
operations.  When processing binary operators in Halide Exprs, it did
not take into account that one of the operands may be non-affine,
resulting in no affine boundary being computed.  The binary combinator
assumed both LHS and RHS bounds exist.  Remove this assumption and
remove an empty affine bound if either of LHS or RHS bound expressions
is not affine.