[C++ API] Rename makeXXXCudaMappingOptions -> makeXXXMappingOptions
This changeset introduces a renaming that will help reduce the delta for
the C++ API refactoring. In particular makeXXXMappingOptions will exist for
different types of mapping options (CPU, CUDA, OpenCL).

The Backend specific MappingOptions functions become generic and not named
after a particular backend.

They can later be used with template types in the overall refactoring.