[C++ API] Loosely typed Duration considered harmful
This commit roots out a cause of regression coming from mixing size_t
with std::chrono and unsafely converting between durations.

The manifestation of the issue was that pruning in the autotuner was not
pruning much affecting overall tuning speed.

There were even 2 declaration of Duration, one in parameters.h and
one in time.h.
This commit tightens up all the loose usages of Duration by basically making it
a new type that has std::chrono::microseconds and only exposes the
minimal API that we need.

We could also drop chrono completely if we don't want to convert between
time intervals.