[C++ API] Extend CpuMappingOptions
This changeset adds supporting functions and objects that will allow an
end-to-end mock CPU example to compile and run, in the context of the global
C++ API refactoring.

This changeset introduces extra methods in CpuMappingOptions to allow it to
be used as a template type with generic function names.

This changeset also moves the implementation of CPU backend specific options
to .cc. This is a general pattern in the refactoring where backend-specific
implementations are fully specialized in .cc files in the proper backend and
templated types and functions live in tc/core and are implemented in -inl.h
files.