Construct MappingFilter from lists of ids and union_pw_affs
MappingFilter node is currently constructed from a union_set, similarly
to a regular Filter node, complemented by a set of mapping identifiers.
Because these identifiers appear in both arguments, MappingFilter
constructor performs presence and uniqueness checks using the list of
CUDA-specific identifiers.  This creates an artificial dependency on
backend-specific code in the otherwise backend-agnostic ScheduleTree
structure.  At the same time, recent changes to MappedScop localized
MappingFilter introduction to a single call that takes a list of affine
expressions to map to identifiers and constructs the filter union_set.

Make the constructor of MappingFilter accept the list of mapping ids and
the list of affine expressions, and move the construction of the filter
union_set inside the MappingFilter constructor.  This breaks the
dependency on the CUDA backend and makes a step towards keeping mapped
affine expressions explicitly in the filter node instead of implicitly
in the set.