[C++ API] Step 10: add the new ATen tensor conversion files
This PR adds the files implementing the ATen user-facing tensor conversion.

This will avoid leaking DLManagedTensors and ScopeGuard at each call site.

As part of the bigger refactoring, these files are not yet activated
(or even compiled) therefore the changeset cannot be tested independently.
This new implementation will be switched in as part of the last commit of the
global refactoring.