Add group normalization benchmark
This commit adds a benchmark and reference implementations for group
normalization (https://arxiv.org/abs/1803.08494).
In particular we found that it was generally beneficial to split the
computation into 2 kernels on GPUs: 1 kernels for the moments and 1
kernel for group normalization from moments. We also provide a
single fused TC which generates a single kernel.