Add a benchmark for first and second moments and sum / sumSquares
In particular this benchmark used to exhibit a performance bug where we
left a lot of performance on the table for on mean / var due to the
normalization update which used to disable reduction detection.
This is fixed in 6fdced39739dd317d20fb811f14a1a084ada96b0

The performance ratio between (sum, mean) or (sumSquares, var) used to be
about 4x on a P100 card.
Note that Max outer fuse followed by Min intra fuse works nicely
and allows reductions to occur.

Lastly, fusing sum and sumSquares only gets an extra 10% over the sum of the
kernels which shows the benefit of fusion.

This has been driven by performance analysis of GroupNormalization which will
be added in a followup commit.