[Bug] Illegal memory access on Volta
This captures a hard to reproduce illegal memory access on Volta.
This seem to happen on Kronecker_1 consistently during tuning.
When try to reproduce I couldn't originally see an issue.
It is only when I put together the last 2 options in a sequence
that I saw a reproducible problem, oly on Volta.
Second set of eyes most welcome here.