Redo the build system
Given the recent developments in PyTorch world and the complexity now
involved in building ATen, it has become counter-productive to build
third-party dependencies from scratch.

Instead, we use conda to get the dependencies for ATen, Caffe2, Halide,
Tapir and PyTorch.

ISL is built from source from the CMakeLists.txt
Google libraries are built and linked statically from a third-party
repository that I created. They are not expected to change, possibly ever.

Static linking of third-party libraries is becoming the de-facto standard
in the conda world. After discussions with @soumith and @yangqing it
seems reasonable for us to follow the same path: avoid dynamic linking
of system libraries. Note that in the upcoming incantation of PyTorch
***even the nvidia libraries will be statically linked***.
This will avoid extremely annoying issues of library conflicts and mismatches
between various components. For us it means that in a near future we will be
able to rely on a fully packaged ATen and Caffe2 that will not bring in shared
libraries in the conda environment.

Immediately however this most recent mode is not yet functional and we still
build a caffe2 conda package on our own. Additionally when installing PyTorch
we need to manually `conda remove cudatoolkit --force` to avoid multiple cuda
libraries from various places. There is not good foolproof solution other than
our dependencies statically linking the world.

On the positive side this commit improves the state of our world as follows:
1. Single unified build for dev and deployment to python packages
2. Build.sh only does one thing (it is still useful to perform some sanity
checks but it only ends up calling cmake for tc)
3. Simplified everything (build, docker files, conda packages, dependencies
management)
4. Anyone should be able to build the TC master easily and use it from C++ or python
5. Build instructions are simplifed from 4 TL;DR pages to a single short one
6. Building from scratch is significantly faster
7. Docker and CI become much simpler and faster
8. No more caches in build .sh

The tradeoffs are:
1. we don't bother with the gcc pre/post 5 ABI issues, we go for GCC5
2. we only officially support Ubuntu16.04 for now with GCC5.4 and Cuda9.0/Cudnn7.1
(other distributions may work too but the compiler/cuda versions are set in stone for now)
3. core team has to use conda

Also, redo Docker + Contbuild

CircleCI builds without cuda even though it does pull the cuda
dependencies for pytorch and caffe2. This can be fixed in the future
once we have the unified pytorch/caffe2 build system and conda packages
made available to us.
CircleCI does not have GPUs and cannot **download them from ether**,
as a consequence we can't test python installation here (our
python bindings require cuda atm) and we only run CPU tests.

The Jenkins build has been reduced to running on a single image.
The process is simplified by avoiding to build a docker image, save it
and download it again.
Instead we just pull an image from dockerhub, update the conda dependencies
then compile and run tests.