MappedScop::detectReductions: use dependences to check validity of ordering
The reduction matching moves out any statements other than
the reduction updates to be able to do a pure mapping
to library calls.
The original code took into account that the init statement
that corresponds to a reduction update statement can be moved
in front of that update statement, but it would punt
on any statements other than those explicitly marked
as init statements.
Use the dependences to check whether moving out the other statements
is valid instead.  This allows the matching to work in case
there are any other statements (that can be moved out) or
if the reduction init statements are not explicitly marked.