extractDomainToIds: do not assume all mapping nodes contain the given ids
A subtree may have mapping filter nodes with different ids.  If this
function is called on such subtree, it would fail even if all branches
of a subtree are mapped to the given ids because it sees a mapping
filter node with different ids.  Simply ignore the node that does not
contain all of the requested ids.

The check was originally introduced in extractDomainToThreads and was
meant to ensure that all branches in the subtree are indeed mapped to
threads.  Instead, perform a stricter check inside extractDomainToIds
that all active domain elements were indeed mapped to the given
identifiers.