memory promotion heuristic: introduce collectMappingsTo
This function returns the union of all mapping filters that map to the
given type of mapping ids.  It is similar in functionality to
extractDomainToIds except for the return type.  In particular, it
ignores elements introduced by extension nodes. It will be used to
collect block and thread mappings in the following commits.