TensorReferenceGroup: replace accessedBySubtree with accessedWithin
This allows for handling block/thread mappings in a more transparent
way.  In particular, computed groups can be made specific to blocks
and/or threads even if they are scoped above the block or thread mapping
by intersecting the domain of the schedule with the mapping filters.

Until now, it was done implicitly by partial schedule computation
functions that intersect the domain of the schedule with the filters
located above it in the tree.  This required promotion functions to be
called below block mapping for shared memory and below thread mapping
for registers.