bandsContainingScheduleDepth: support depth 0
Technically, the first band in each subtree contains depth 0.  However,
the check currently implemented in bandsContainingScheduleDepth expects
the depth before the given band to be strictly less than the given
depth.  Furthermore, the subsequent uses of this function will attempt
to split the resulting bands so that the required depth corresponds to
the last member of the band.  It is impossible to split off all band
members.  The result the callers of bandsContainingScheduleDepth
followed by splitting expect is a band with zero members to be used as a
promotion scope.  Add a special case to bandsContainingScheduleDepth
that introduces such zero-dimensional band below the domain node (or the
context node immediately following the domain) and returns it.
Subsequent band splitting will detect that zero-depth is reached at the
end of the newly introduced band and will stop there.