register promotion: promote groups with more than one element
Relax the restriction for a promoted group to contain only one element.
Instead, check that all only the dimensions that either correspond to
unrolled loops or are mapped to blocks or threads appear in the access
subscripts.  During code generation, the corresponding loop iterators
are either replaced by constants due to unrolling or become implicit in
block/thread-specific groups (i.e., each thread in each block accesses
different elements so it can use the same "address" for them).