use polyhedral reduction description to detect reductions
The exact notion of what constitutes a "reduction" was
not spelled out very clearly, but in the use case
of mapping the detected reduction to CUB, each thread
needs to write to a single tensor element.
Take this property as the definition of a "reduction".

A necessary and sufficient condition is that
the schedule dimensions outside the band members
that are considered to form the reduction,
determine a unique tensor element that is updated.

The condition used in the original detection mechanism
(that the reduction band members only involve a specific
reduction dimension of the reduction update statement)
is neither sufficient nor necessary.
It is not sufficient because it doesn't force
the schedule dimensions not involving the reduction dimensions
to appear outside of those band members.
It is not clear if such a case could ever arise in practice
in TC, because the coincident band members get placed outermost.
However, this situation could change in the future if other
criteria such as consecutivity get taken into account.

It is not necessary because it is perfectly fine
for such a band member to reference multiple
reduction dimensions or even to involve non-reduction
dimensions, as long as those non-reduction dimensions
also appear in outer schedule band members.
Since no skewing is currently allowed, this
is currently also unlikely to happen.

In summary, what the reduction band members look like
is largely irrelevant.  What is important is that
the outer band members make sure that only a single
tensor element gets accessed.  It just so happens
that the reduction band members are very likely
to have a particular form in that case.

Change the reduction detection mechanism to use
the polyhedral reduction representation.
Note that if a tensor element is involved in multiple
separate reductions (again, unlikely in the current
code base), then the detection will not only ensure
that a single tensor element is involved, but more
specifically, a single reduction on that tensor element.

This new mechanism is also much simpler.